+++ homebrew/Views/BrewDetail.swift
@@ -192,46 +192,3 @@ struct BrewDetail: View {
         refreshID = UUID()
     }
 }
-
-func getAge(of brew: Brew, fromBottlingDate: Bool = false, fromSecondary: Bool = false) -> Double {
-    var from = brew.startDate
-    var to = Date()
-    if fromBottlingDate {
-        from = brew.bottles?.date
-    } else if fromSecondary {
-        from = brew.secondary?.date
-        if let bottle = brew.bottles {
-            to = bottle.date ?? Date()
-        }
-    } else {
-        if let seconary = brew.secondary {
-            to = seconary.date ?? Date()
-        } else if let bottle = brew.bottles {
-            to = bottle.date ?? Date()
-        }
-    }
-    let brewAge = to.timeIntervalSince(from ?? Date()) / 86400
-    return brewAge
-}
-
-enum BrewStage: String {
-    case primary = "Primary Fermentation"
-    case secondary = "Secondary Fermentation"
-    case bottled = "Bottled"
-}
-
-func getStage(of brew: Brew) -> BrewStage {
-    if brew.bottles != nil {
-        return .bottled
-    } else if brew.secondary != nil {
-        return .secondary
-    }
-    return .primary
-}
-
-private let brewDateFormatter: DateFormatter = {
-    let formatter = DateFormatter()
-    formatter.dateStyle = .short
-    formatter.timeStyle = .short
-    return formatter
-}()
